+++ example/urls.py
@@ -4,13 +4,13 @@ from django.conf.urls import url
 from django.urls import include, path, re_path
 from django.conf.urls.static import static
 from django.contrib import admin
+from pages import views
 
-admin.autodiscover()
-
+from pages.urlconf_registry import register_urlconf
+register_urlconf('blog', 'blog.urls', label='Blog index')
 
 urlpatterns = [
     url(r'^i18n/', include('django.conf.urls.i18n')),
-    # url(r'^grappelli/', include('grappelli.urls')),  # grappelli URLS
     url(r'^admin/', admin.site.urls),
 ]
 
@@ -20,16 +20,12 @@ try:
 except ImportError:
 	pass
 
-urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT, show_indexes=True)
-urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT, show_indexes=True)
-
-
-if 'ckeditor' in settings.INSTALLED_APPS:
-    urlpatterns += [
-        url(r'^ckeditor/', include('ckeditor.urls')),
-    ]
+urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
 
 urlpatterns += [
-    # this will match /admin if someone forget the traling slash
+    url(r'^$', views.details, {'path': '', 'name': 'pages-root'}),
     url(r'^', include('pages.urls')),
 ]
+
+admin.site.site_header = 'Gerbi Admin'